hf_service/requirements.txt
fastapi
uvicorn
httpx
pillow
torch
transformers
